HOLO-142 <refactor: 리팩토링 및 메타데이터 추가>
