+++ src/app/policy/_components/PolicyResult.tsx
@@ -1,5 +1,5 @@
+import PolicyResultCard from "@/app/policy/_components/PolicyResultCard";
 import type { PolicyData } from "@/app/policy/_types/policy";
-import Link from "next/link";
 
 type PolicyResultProps = {
   error: Error | null;
@@ -8,22 +8,23 @@ type PolicyResultProps = {
 };
 
 const PolicyResult = ({ error, isFetching, policyData }: PolicyResultProps) => {
-  if (error instanceof Error) return <div>에러: {error.message}</div>; //TODO: 에러 변경
+  if (error instanceof Error) {
+    return <div className="text-red-500">에러: {error.message}</div>;
+  }
+
+  if (isFetching) {
+    return <div>조회 중...</div>;
+  }
+
+  if (!policyData) {
+    return <div>검색 결과가 없습니다.</div>;
+  }
 
   return (
     <div className="grid gap-4">
-      {isFetching && <div>조회 중...</div>}
-      {!isFetching &&
-        policyData &&
-        policyData.map((policy) => (
-          <Link
-            href={`/policy/${policy.bizId}`}
-            key={policy.bizId}
-            className="cursor-pointer rounded-lg border p-4 hover:bg-gray-50"
-          >
-            <h3>{policy.polyBizSjnm}</h3>
-          </Link>
-        ))}
+      {policyData.map((policy) => (
+        <PolicyResultCard key={policy.bizId} policy={policy} />
+      ))}
     </div>
   );
 };
