+++ src/app/policy/error.jsx
@@ -1,0 +1,1 @@
+"use client";
