+++ src/app/policy/page.tsx
@@ -1,4 +1,10 @@
 import PolicyCont from "@/app/policy/_components/PolicyCont";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "청년 정책",
+  description: "청년들을 위한 정책 제공",
+};
 
 const page = () => {
   return (
